ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.2
scikit-learn==1.5.2
xgboost==2.1.3
pandas==2.2.3
numpy==2.1.3
joblib==1.4.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
